refactor: initial setup
